Update index.md
removint text in the danger zone: This is the workshop template. Delete these lines and use it to customize your own website. If you are running a self-organized workshop or have not put in a workshop request yet, please also fill in this workshop request form to let us know about your workshop and our administrator may contact you if we need any extra information. If this is a pilot workshop for a new lesson, set the `pilot` field to `true` in `_config.yml`. For workshops teaching a lesson in The Carpentries Incubator, remember to uncomment the `incubator_lesson_site`, `incubator_pre_survey`, and `incubator_post_survey` fields in `_config.yml`